refactor: based in a Notifier
