+++ crates/observable/src/observable.rs
@@ -1,115 +1,10 @@
 use std::{cell::Ref, cell::RefCell, rc::Rc};
 
-struct ListenerSet<T> {
-    nextid: usize,
-    items: Vec<ListenerItem<T>>,
-}
-
-impl<T> Default for ListenerSet<T> {
-    fn default() -> Self {
-        ListenerSet {
-            nextid: 0,
-            items: Vec::new(),
-        }
-    }
-}
-
-struct ListenerItem<T> {
-    /// Monotonic id for use in the binary search
-    id: usize,
-    listener: Listener<T>,
-}
-
-impl<T> ListenerSet<T> {
-    fn working_set(&mut self) -> Vec<Listener<T>> {
-        // It's possible to add listeners while we are firing a listener
-        // so we need to make a copy of the listeners vec so we're not mutating it while calling listener functions
-        let mut working_set: Vec<Listener<T>> = Vec::with_capacity(self.items.len());
-
-        let first_once_pos = 'first_once_pos: {
-            for (index, item) in self.items.iter().enumerate() {
-                match &item.listener {
-                    Listener::Once(_) => break 'first_once_pos Some(index),
-                    Listener::Durable(f) => {
-                        working_set.push(Listener::Durable(f.clone()));
-                    }
-                }
-            }
-            None
-        };
-
-        // only moves durables if necessary while fills the working_set
-        if let Some(first_once_pos) = first_once_pos {
-            let items = unsafe {
-                let items = &mut self.items as *mut Vec<ListenerItem<T>>;
-                (*items).drain(first_once_pos..)
-            };
-
-            for item in items {
-                match &item.listener {
-                    Listener::Once(_) => working_set.push(item.listener),
-                    Listener::Durable(f) => {
-                        working_set.push(Listener::Durable(f.clone()));
-                        self.items.push(item);
-                    }
-                }
-            }
-        }
-
-        working_set
-    }
-
-    fn notify(&mut self, value: &T) {
-        let working_set = self.working_set();
-
-        // Now that the borrow on the listeners vec is over, we can safely call them
-        // We can also be confident that we won't call any listeners which were attached during our dispatch
-        for listener in working_set {
-            match listener {
-                Listener::Once(f) => f(value),
-                Listener::Durable(f) => f(value),
-            }
-        }
-    }
-
-    fn _subscribe(&mut self, listener: Listener<T>) -> ListenerHandle {
-        let id = self.nextid;
-        self.nextid += 1;
-        self.items.push(ListenerItem { id, listener });
-        ListenerHandle(id)
-    }
-
-    pub fn subscribe(&mut self, cb: Box<dyn Fn(&T)>) -> ListenerHandle {
-        let listener = Listener::Durable(cb.into());
-        self._subscribe(listener)
-    }
-    pub fn once(&mut self, cb: Box<dyn Fn(&T)>) -> ListenerHandle {
-        let listener = Listener::Once(cb);
-        self._subscribe(listener)
-    }
-
-    pub fn unsubscribe(&mut self, handle: ListenerHandle) -> bool {
-        // Find the current listener offset
-        match self.items.binary_search_by(|probe| probe.id.cmp(&handle.0)) {
-            Ok(offset) => {
-                self.items.remove(offset);
-                true
-            }
-            Err(_) => false,
-        }
-    }
-}
-
-pub enum Listener<T> {
-    Once(Box<dyn Fn(&T)>),
-    Durable(Rc<dyn Fn(&T)>),
-}
-
-pub struct ListenerHandle(usize);
+use crate::notifier::{ListenerHandle, Notifier};
 
 pub struct Observable<T> {
     value: Rc<RefCell<T>>,
-    listener_set: Rc<RefCell<ListenerSet<T>>>,
+    notifier: Rc<Notifier<T>>,
 }
 
 // Implemented manually because `T` does not need to be Clone
@@ -117,7 +12,7 @@ impl<T> Clone for Observable<T> {
     fn clone(&self) -> Self {
         Observable {
             value: self.value.clone(),
-            listener_set: self.listener_set.clone(),
+            notifier: self.notifier.clone(),
         }
     }
 }
@@ -126,7 +21,7 @@ impl<T> Observable<T> {
     pub fn new(value: T) -> Self {
         Self {
             value: Rc::new(RefCell::new(value)),
-            listener_set: Rc::new(RefCell::new(ListenerSet::default())),
+            notifier: Rc::default(),
         }
     }
 
@@ -136,7 +31,7 @@ impl<T> Observable<T> {
             *(self.value.borrow_mut()) = value;
         };
         let r = self.value.borrow();
-        self.listener_set.borrow_mut().notify(&r);
+        self.notifier.notify(&r);
     }
     // }
     // impl<T> Observe<T> for Observable<T> {
@@ -145,14 +40,14 @@ impl<T> Observable<T> {
     }
 
     pub fn subscribe(&self, cb: Box<dyn Fn(&T)>) -> ListenerHandle {
-        self.listener_set.borrow_mut().subscribe(cb)
+        self.notifier.subscribe(cb)
     }
     pub fn once(&self, cb: Box<dyn Fn(&T)>) -> ListenerHandle {
-        self.listener_set.borrow_mut().once(cb)
+        self.notifier.once(cb)
     }
 
     pub fn unsubscribe(&self, handle: ListenerHandle) -> bool {
-        self.listener_set.borrow_mut().unsubscribe(handle)
+        self.notifier.unsubscribe(handle)
     }
 }
 
@@ -167,7 +62,7 @@ where
             vec.push(item);
         }
 
-        self.listener_set.borrow_mut().notify(&self.value.borrow());
+        self.notifier.notify(&self.value.borrow());
     }
 }
